requirements.txt
samtools
Biopython
pandas
numpy
sklearn
torch
progressbar2
networkx
(community)
